+++ app/changelog/changelog.ts
@@ -2,7 +2,11 @@ export const changelog: ChangelogItem[] = [
   {
     version: '1.13.0',
     type: 'minor',
-    added: ['Posts können nun über das Kebap-Menü gemeldet werden.'],
+    added: [
+      'Posts können nun über das Kebap-Menü gemeldet werden.',
+      'Private Nachrichten können nun als ungelesen markiert werden.',
+      'Private Nachrichten können nun in einen anderen Ordner verschoben werden.',
+    ],
   },
   {
     version: '1.12.0',